+++ app/(route)/(auth)/signup/_components/sign-up-form.tsx
@@ -5,7 +5,7 @@ import type { FieldError, RegisterOptions } from 'react-hook-form';
 import { useForm } from 'react-hook-form';
 import { useRouter } from 'next/navigation';
 import type { SignupFormValues } from '@/_apis/type';
-import { postSignup } from '@/_apis/user';
+import { postSignup } from '@/_apis/users';
 
 import { useModal } from '@/_hooks/useModal';
